livepatch: alert when snapd does not have wait cmd
When enablig livepatch, we run a snapd wait operation before
installing the canonical-livepatch snap. On older system that did
not upgrade snapd, the wait command may not be available. In that
situation, enabling livepatch will fail. We are now updating the code
to alert the user about this situation, but we proceed with enabling
livepatch.

LP: #1927905