enable: don't allow disabling services on beta check
If we fail to enable a service, we call the can_enable method directly.
We do that to verify if the reason we failed to enable a service is
because the service is in beta mode. However, the can_enable method
calls the handle_incompatible_services method, which may prompt users
to disable the incompatible service again. Since this check is only used
to see if the service is beta, we should not allow for disabling
incompatible services when we do it.

Fixes: #1652