repo: allow downgrading packages during enable
When enabling a repo entitlement, like FIPS, we are currently not
allowing for package downgrades. This is problematic if users are
using package from a third party, that could have versions higher
than the ones we are delivering. To enforce that users will run
the right packages for the services we are enabling, we are now
supporting downgrading packages when enabling a service.

Fixes: #1659