tests: remove clashing lxd user-data
lxd relies on a write_files vendor-data directive to set up the
lxd-agent on lxd vms of some ubuntu releases. Our use of write_files was
overwriting the vendor-data and preventing the lxd-agent from coming up.

Here we use a runcmd to accomplish the same thing we were using
write_files for in our integration tests.