refresh: add target to refresh command
ua refresh now accepts an optional arg, which can be config or contract.
Depending on the value, ua refresh will either process any config
changes or re-fetch the contract from the server. If the arg is not
provided, then ua refresh will do both.

ua refresh config will also print a warning if there is no proxy set in
uaclient.conf but there is a snap or livepatch proxy set. We can't
actually unset a snap/livepatch proxy in this scenario because this may
have been intended by the user. Instead of unsetting, we provide
instructions to either remove the snap/livepatch proxy setting, or add
the proxy config to uaclient.conf.