status: do not modify internal status state
We have modified our internal status state to better reflect
the contract structure we want to output in the json format. However,
by modifying the internal we end up affecting non-root users. That is
because when a non-root user runs ua status, we get the status directly
from the file saved on disk. If the user has not run any root commands
after installing a new version of uaclient, that file on disk will be
stale and will not be able to be used by the status formatting code.
Because of that, we are bringing back the original internal status
representation.