config: remove config_path from config dict
We are now removing the config_path entry from the
config dict and pushing that logic into a dedicated function.
We are using that to not clash with the write_cfg method
which by default writes all content of the config dict into disk.
Since the config_path is a variable that is being used only
internally, we don't need it on disk.