fips: do not block enable on conditional packages
When enabling fips we install FIPS version of some packages
if they are already installed in the system. Currently, we
block enable if we can't install any of those conditional packages.
Now, we are alerting the user about errors on installing those
packages, but we don't block the enable operation