tests: build packages using sbuild
When running integration tests with BUILD_PR=1, we build the ua package
using in the same instance type that we are going to run the test on.
This is not ideal, since we run integrations tests on clouds and LXD
vms. We are now building the packages through sbuild, which is a faster
and more correct approach for building the packages to be used in the
test