tests: don't run apt update on userdata
When installing ua directly from proposed, we create an userdata
that create the proposed source file and runs apt update. This can
cause race conditions on the start up process of an instance, since
another apt command might be running. To try to avoid that, we are
now running the apt update command only when we really need it