tests: drop cis check on focal to groovy upgrade
When running the focal to groovy upgrade test, we check
ua status after successfully upgrading to groovy. In that check,
we verify if cis is present in the output. However, during upgrade
we can end up in a groovy ua package where cis is still in beta and
will not show up in status. Therefore, we are dropping that line
on the check