tests: default to installing ua for each scenario
The strategy of creating a snapshot with ua installed before running any
integration test has a significant upfront cost. There are cases, such
as local development on one feature, where this upfront cost is not
worth it. This commit adds an option so that the snapshot strategy with
upfront cost can be optionally turned on, but defaults to off.