cloud: rely only on cloud-init to determine cloud and instance IDs
This change discards fallback files when looking for instance ID or
cloud ID. If cloud-init is present on the system, it will handle those
informations through the instance query and cloud-id commands.

If the detection through cloud-init calls does not work, we assume not
on cloud. This solves the scenarios where there is a failure because
cloud-id is not there, but the fallback files are, and they don't have
the expected data.

LP: #1940131

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>