ros: update message for required/dependent service
When enabling ROS, we might need to enable required services first.
When that happens, we want the enable opearion for these required
services to be silent. This also happens on disable, when we are
disabling a service ROS depends on. We are moking the enable/disable
of those required/dependent services silent