fix: do not fail if non-utf8 locale is being used
When running an ua fix command, we tell at the end if the
issue is resolved or not. In that message, we have some unicode
characters that are displayed. If the user is running on a system
that doesn't support those characters, the printing that last message
will fail. We are now redacting the code to not fail under this
condition.

Fixes: #1463