postinst: remove old ua-messaging.timer services
After upgrading the package we can completely remove the
ua-messaging.timer from the system. Currently, the postrm script
will mask this service, meaning that the service is inactive but
will still appear in the systemctl list-timers --all command.
By removing the symlink created by the mask operation and reloading
the systemd daemon, we will guarantee that the service will not
be found anymore.