timer: license_check only runs on gcpgeneric
We've moved the 5 minute timer to a separate systemd unit that is
activated via a marker file. This allows us to only activate it on
relevant platforms, which is currently only GCP instances on an LTS
release that are not attached. We create the marker file in our postinst
script if necessary, and will seed it on appropriate GCP LTS images.

With that high frequency job separated, we now drastically reduce the
frequency of the timer that runs on all ubuntu instances.

To manually disable the 5 minute timer on a gcp instance, run the
following:
```
sudo systemctl stop ua-license-check.timer
sudo rm /var/lib/ubuntu-advantage/marker-license-check
sudo systemctl stop ua-license-check.timer
```
Note that the "stop" command must be run before and after the "rm"
command because some systemd versions will get confused otherwise.