metering: update contract response structure
Currently, the machine activity endpoint will not return the activity
data inside using a structure where activityInfo is the top level key.
However, this structure will be used in the response of attach and
refresh responses. We are adapting the machine activity endpoint
to be consistent with the other responses