fix: check reboot-required.pkgs to reboot
During some UA operations we may install some package on the system.
Those packages may require a reboot to complete their setup.
Currently, we only check the existence of the /var/run/reboot-required
file to notify the user that the system needs a reboot. However,
that marker file could have been generated by a different package
install outside of the UA operation. To address that, we are now
using the /var/run/reboot-required.pkgs file to check if the packages
installed during the UA operation are on that file. Only if we have the packages
there together with the reboot marker file, we will ask the user to reboot
the system.