postinst: make check for gce instance more robust
Currently, we have a check on postinst to verify
if we are on a gce cloud. We only verify if the gce
string is in the cloud-id of the instance. Therefore,
we are open to matching non-gce clouds that have the gce string
on their cloud-id. We are now updating this part of the code
to make the match more robust by verifying if the cloud-id starts
if the gce string.