repo: update apt url check for commented source
When we are running the process_contract_deltas function,
we verify if the delta apt url is already on the service
apt source file. However, there are situations were there
will no apt url deltas and we still want to regenerate the service
apt source file, for example if we have a commented apt source file.
We are updating the code to address that situation.