contract: save machine id during attach operation
When running an attach operation we are now saving the machine-id
into our cache. We need to have that information saved on the cache
to properly perform the detach operation. We are now updating our
contract logic to save that information after we receive the machine
token during an attach operation