postinst: fix service check for unattached machine
When running in a LTS machine with an architecture that is different from
i386 and amd64 we will try to unconfigure esm services on these
non-supported architectures. Before we unconfigure those services, we
check to see if they are enabled. However, the code that performs that
check is looking for a key in the status cache that is only present
when machine is attached to a subscription. We are now updating the
code to not fail if the machine is unattached.

LP: #1951705
Fixes: #1888