tox: set version of setuptools for some tests
For some older ubuntu releases, we need to set up an older version
of pip. That's because the constraints we have for the older releases
could not be successfully used with new versions of pip. This pip
restriction is now affecting new versions of setuptools as well. Because
of that, we are now setting up the versions of setuptools for those
older ubuntu version tests