daemon: cloud-id file shim
This introduces a new systemd service that will run on startup only if
the version of cloud-init does not create the cloud-id file. The service
creates the same cloud-id file that cloud-init does in newer versions.
This is specifically necessary for xenial, and is a noop if cloud-init
is new enough.

It also runs the same script on postinst to cover the scenario where the
new version of cloud-init is installed but the instance has not
rebooted, and so cloud-init hasn't yet created the cloud-id file.