CI: restrict the cloud-cleanup jobs to the base repository
In forks there won't be secrets, so the cleanup jobs will fail.
I tried checking if the secret are available instead of checking
the repository name, with something like:

  if: ${{ secrets.SOMESECRET != '' }}

but this can't be done because of:

  https://github.com/actions/runner/issues/520

There are ways to check if secrets are available using helper jobs,
but I don't think it's worth it for the moment.