cli: add refresh subcommand to update motd message
Right now, users have no direct way to update the MOTD messages
UA deliver. For example, if a contract is renewed, users will need
to wait for our timer job to run so they can see the expiry message
on MOTD disappear. We are now adding a `ua refresh motd` command
that updates the MOTD message