apt: allow caching apt-cache policy commands
We currently use the output to check if each repo entitlement
is enabled in the system. This happens for example when we are
running an ua status operation, where we run an apt-cache policy
for each entitled repo service the user has. We can cache that command
if no update calls are performed. Therefore, we are speeding up
the execution of ua status by caching the result of apt-cache policy
calls.