enable: fix apt auth line replacement
The regex was failing to grab the entire url and would only match
"esm.ubuntu.com" for each line. Because every service is a more specific
path of esm.ubuntu.com and there was no check to prevent multiple
additions of a new line, then a new line would be inserted before every
existing line each time.

This changes to a simpler split-based approach to get each line's url
and compare. We also add an integration test to insure the behavior is
not inflating the size of the file.

LP: #1985863