status: make lock world readable so non-root status --wait works
changes to get_machine_id are because it was previously relying on the
lock file creation to create the private subdirectory. Now that the lock
is not in the private subdir, that directory wasn't in place and
creating the machine-id file failed. Using the `cfg.write_cache` method
handles creating the private subdirectory.

The exist_ok was added in write_cache because of a race condition in the
unit tests where that line was reached and the directory already
existed.