tests: avoid aws debug messages when testing
The AWS boto libraries are very very very verbose.

We pretty much never want their debug level messages,
but we do want to be able to use the debug log level
in our code. So we bump their loggers up to info
when we set debug at the cli with --logging-level=debug