api: best_effort full_auto_attach instead of fail fast
full_auto_attach.v1 will no longer raise exceptions about invalid
arguments before attempting to attach. Instead it will attempt to auto
attach and enable all services. If any services fail, they will all be
reported in a single error.

This also adds a check for the disable_auto_attach configuration and
raises an error if that is true before the attempt to attach.

This is technically a breaking change, but because we caught this early
and we know of the only current consumer of this function we can make it
safely without bumping the version.