auto-attach: use api in cli entry-point
In order to simplify the code base, make use of
full_auto_attach api endpoint in cli `pro auto-attach` command.

Behavior change:

Previously, in the case where the user tried to auto attach using
the cli `pro auto-attach` on instances that were already attached
or have disable_auto_attach configured as true in the
uaclient.conf file, the cli exited with a 0 exit_code.

Now, the cli returns non 0 exit codes in both cases. The cli returned
0s because we would call the cli command on the
ua-auto-attach.service systemd unit. If we detected that the user 
was already attached or hand explicit config disabling that service,
we didn't want to show that this service failed.

After previous and this refactoring, both entry-points make use of the
new external api to auto-attach (full_auto_attach) and handle the error
cases differently.

We didn't advertise the use of `pro auto-attach` for programmatic
consumption.

Therefore we believe it is safe to change the exit codes in these
scenarios.