chore: do not import version on setup.py
It is worth nothing, and sows circular dependencies for us to reap.

The package is installed on Ubuntu using apt, and all of the machinery is in place to properly version the package when building (whether locally or on PPAs) and report it appropriately.

The problem with setup.py importing from version brings the fact that version itself imports other modules, and:
a) There is no guarantee that those modules will not import version again, generating a circular dependency
b) There are guaranteed dependencies from the package perspective (yaml, apt) which break when importing in setup.py because no requirement installation was done yet, if any.

The only tooling that relies on installing through setup.py is tox, for the tests, and it can be any version - tox is happy with whatever we set.

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>