motd: refactor contract status message delivery
Still use timer because ensuring the timely and accurate delivery of
this message requires a network request, which shouldn't be done at
motd-display time.

Calculate apt packages in timer instead of via hook. This means the
count could become out of date by a few hours, but we consider this
acceptable. This allows us to delete the old apt hook.