contract: send host_type when getting resources
When unattached, we send platform information to the contract server
that replies with which services are available. One gap that we have is
that some services are not available on all host types. For example,
Livepatch cannot be enabled on a container. We are now sending that
information to the contract server so that they can take that
information in consideration when informing which services are
available.