Add check to not prompt user on apt_news changes
In the previous release we have changed the uaclient.conf file
to capitalized a word. If the user has changed apt_news through pro config,
the user will now get a debconf prompt because of it. This is not ideal,
since we are the ones that instruct the user to run that command if they
don't want apt advertisement on their system. This PR is creating a
mechanism to avoid that prompt if the has just changed the apt_news
config