config: migrate user config out of uaclient.conf
uaclient.conf is a conffile and as such gets special treatment whenever
the user changes it and when we change the default file. We don't want
this special treatment for several of the fields currently in
uaclient.conf, all in the sub-document: "ua_config". These fields are
the ones directly altered via the cli command `pro config set`.

To avoid issues with settings modified by `pro config set` causing
unnecessary conffile prompts, we are moving the entire "ua_config"
sub-document out of uaclient.conf and into its own file located in
/var/lib. This file will now be exclusively managed by the `pro config`
commands.

The major part of this commit is the migration preparation in preinst,
and the migrate_user_config.py script that is run in postinst. These are
designed to avoid unnecessary conffile prompts while maintaining all user
settings.

LP: #2004280