system: use VERSION_CODENAME and VERSION_ID
When parsing the VERSION field from /etc/os-release, we might
not have the series in the VERSION string. Because of that, we
are now relying on both VERSION_ID and VERSION_CODENAME to
extract release info from. Only if we fail to the that information
from those field we will fallback on parsing VERSION