fix: verify if package can be installed
When running a pro fix command, we check if the fixed package version
is higher than the one currently installed. If it is, we try to update
the package, but we don't check if the package can indeed be upgrade.
For example, if the version that fix the error is higher than the
package candidate version, we cannot install that fix. To address that,
We are now checking the target package candidate version to address
those potential scenarios

LP: #2006705