ci: remove old python version unit-tests job
The unit tests and flake8 already run against the correct python version
as part of the package build jobs.

We still run the unit tests and flake8 on the latest version of python
as a gate to prevent CI from moving on to package builds unnecessarily.

Since we're not doing the special python version work, I moved flake8
into the "Static Analysis" job. I kept unit tests separate to easily
distinguish if they failed via the GH actions UI.