apt-cache: only setup the local cache if the service is available
Some users would see the esm available updates even if they could not
enable esm, due to being running on some unsupported architecture, for
instance.
Now we look into the availability of the service for a given system
before setting up the cache, and ensure no cache is there if a service
is not available.

LP: #2004018

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>