security_status: better isolate apt.Cache
There is problem on the apt library when we import and use
both apt_pkg and apt in the same module. For example, once we
create the esm cache, that affects the apt.Cache creation. To avoid
that, we are now better isolating the apt.Cache creation to handle
potential configuration changes caused by the esm cache