fix: separate target USN from related USNs
When we fix a USN, we also try to fix related USNs. However,
in that setting, if we failed to fix a related USN we would say
that the requested USN could not be fixed, which is incorrect. We
are now separating the requested from the related USNs when the user
is running pro fix