apt: restore apt configuration after creating the esm cache
When the esm cache is created, the apt_pkg.config globals are set to
values a regular user of the library wouldn't expect.
We need to set it back to the defaults, by clearing the config and
calling init() again, so any subsequent call of apt.* will read the
globals the same way it would without our interference.

LP: #2008280

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>