yaml: catch and surface errors with the yaml installation
When working on the fix of LP #2007241 and LP #2007234, some logic was
implemented to import only the system-installed yaml package. However,
after further discussion, the team, with the help of reviewers,
concluded that this is indeed an error, and should be just treated
correctly to avoid a crash (and consequently an oops report).

This commit substitutes the previous logic with two exception handlers:
one for the missing yaml package, another for any yaml package which
cannot perform the basic operations we need.

In both cases, we log an error and exit 1.

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>