docs: clarify location of apt_check.py
The hint we are currently giving suggests it would be from the
update-notifier package while in reality from a users POV the
package it is in is actually update-notifier-common. Since end
users usually do not care much about source package names, but
know binary packages we should clarify that.

Furthermore apt_check.py out of that package is called a script,
but it is not in a binary path and hence people can not just call
it like other programs. Therefore on the first occurance we should
list the full path.