docs: add an example of an unattached machine
We have missed the output of an unattached machine.
The output examples we already had are mostly seen by those using
pro consciously, but we should have search engines also (or especially)
pick up the strings used when being unattached as that might be
the most likely kind of people looking for documentation what
it is about.