docs: clarify update notifiers task and invocation
The esm mentioning might make it appear new to some, but other
than differentiating the additional potential sources of updates
it has always done that. Mention this in the section about
update-notifier.

Furthermore we mention the script, but if someone just calls it
they get a very disatisfying "0;0" as the return. Hint the reader
at `--human-readable` to recognize the output they see in MOTD.