docs: explain both MOTD interactions
We have given examples of the content created by update-notifier
and that about what pro generates. But then in the "how does it work"
section we suddenly only talk about one of them - the one that all
unattached people will never see.

Also mention that and split the headers in that section to match what
we use in the description to be recognizable.