docs: mention the esm announcement message
While we have just removed it, people are still seeing it and will
find it in old logs and posts about the overall topic.
Let us explain what it is, why it was added (and removed).

Just like the others this has a section how it looks like and one how it
is generated/inserted.

Fixes: #2533