Add XML sitemap to docs
Add XML sitemap for Google Search Console

GSC has not been able to correctly find more than 12 pages in
our docs because the default RTD sitemap is pointing to a URL
Google doesn't expect.

Adding a new sitemap specifically for GSC so that the docs
can be crawled and then reliably listed in search results.