livepatch: support new version of supported kernels api
The behavior of the livepatch API endpoint for querying kernel support
is changing without bumping the version. The new version is currently on
staging and the old version is blocked from going to production by a
bug affecting infrastructure. Because we don't know exactly when the new
version will reach production, but we have an impending scheduled
release, this commit supports both the old and new version at the same
time.