refactor: implementation-agnostic interface to readurl
The previous interface to readurl had different return types depending
on the request, and it propogated exceptions that were raised from the
urllib library. Those urllib exceptions were then caught and used in
various places throughout the codebase.

The upcoming tls-in-tls implementation will require swapping out the
urllib request function for a pycurl-based one, which will not have any
of the same return types or exceptions as urllib, so we need to stop
using urllib-specific things everywhere.

To achieve this, I've defined a simple HTTPResponse tuple that will
always be returned by the readurl function - this should be easily
implemented by any underlying http implementation. Notice that the
function always returns this tuple, even if there is a http code above
300. Previously we expected/excepted errors on 4xx/5xx responses, so
I've changed a lot of the calling code to check the return code rather
than except errors. The only error that is raised is now our
exceptions.UrlError and I've redefined it to only be raised if something
about the connection fails.