docs: Add sync tabs to API reference (#2617)
The API reference document showed both Python and
CLI code to the user at the same time. Now through
the use of tabs, the user can choose whether to view
Python or the CLI version. Tabs are synced so their
choice is applied to all entries. The API page is
now made cleaner and more functional.

This also converts the source doc from md to rst