refactor: replace confusing env arg to subp
The only reason the env argument was used was for inserting values that
were intended to be additional environment variables on top of the
current process' environment. There were also issues with modifying
default arg values and modifying caller-provided dicts to the env
argument.

To address this, I'm renaming the argument to make it clear it is
treated differently than the stdlib's "env" argument to various
subprocess functions. I've also removed all related default arg values
of {} to avoid accidentally modifying them. And I've changed the
use of the new argument to correctly extend os.environ without
modifying caller state.