anbox: ignore enableByDefault obligation
The anbox product has the enableByDefault obligation turned on because
the old setup of Anbox require the resource tokens to be already
present in the machine-token file once the service is enabled.
However, we don't need to enable this service during attach.
Finally, we should remove this code once the Contract server provides
another mechanism to deliver resource tokens by default to some
services