systemd: change u-a.service type from notify to exec
We have a dependency on python3-systemd to use the 'notify' function,
which cannot be satisfied as the package is not part of ubuntu-minimal.
This dependency was never declared in the debian/control file anyway,
and we are removing it.

We don't lose much in the change anyway, as we are only configuring
logging before notifying.

Fixes: #2692
LP: #2038417

Signed-off-by: Renan Rodrigo <renanrodrigo@canonical.com>