config: unify conflicting PRIVATE_SUBDIR definitions
We had conflicting definitions of the same name PRIVATE_SUBDIR and
due to that some places using it in one or the other way from one
of the two sources.

Furthermore the definitions were different, once with and once without
a leading slash.

This commit unifies this to one place defining the string in just one way
(no leading slash) and adapts users that assumed differently.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>