tests: add scenario where cloud-init is disabled
Add a scenario where cloud-init is disabled after boot.
In that situation there is no cloud-id command anymore. To test
that the Pro client doesn't break, we will enable FIPS on a cloud
instance to show that service is indeed enabled even though
the cloud metapackage is not installed

LP: #1938208