fix: remove redundant message for enable operation
We we require an enable operation for the fix command,
we are now displaying two messages if the enable part fails.
One related to the actual error and a generic one mentioning
that we need the service to install the packages. We are removing
the generic one as it does add any new information to the user.