Move temporary test files to /var/lib/ubuntu-advantage
Due to the PrivateTmp=yes restriction set on the apt-news.service unit,
tests that use the /tmp directory will fail.

This happens because the unit will get a brand new and empty /tmp when
it starts up, and anything it does in that /tmp will be gone when it
terminates.