fips-preview: remove cloud specific checks
On GCP, we block enabling FIPS if we detect that
there is no FIPS cloud optimized kernel. However, since
fips-preview is intended for early access and the packages
there are not yet certified, we believe we should not block
enabling the service on clouds. Additionally, fips-preview
is already using default fips metapackages for other clouds
we support