tests: implement behave-any for all tests
This extends the "behave-any" refactor to all integration tests. That means any combination of tests, regardless of the cloud/platform they run on, can be run with one command.  This allows consolidation of some tests since tests can now be parametrized by machine_type.

There is still the capability to run the tests in a per release, per platform matrix via the run-integration-tests.py script. However, after this refactor we don't have to split the tests up that way.