update logic that checks if a service is enabled
When enabling a repo service, we check to see if it apt url is present
at the output of apt-cache policy. We are updating this check to also
check if any suite of the service also appear together with the apt url.
We need this for some setups where the service share the same apt url.

LP: #2031192