docs: split fix intro and scenarios
As of today there is only a tutorial, but for a tutorial it feels
overloaded with special cases due to the many scenarios.

On the other hand we have no good "how to understand all those scenarios
that does not start with multipass and guest creation".

Both would IMHO be better than what we have, so let us split and cross
link the two as needed. Shared content is pushed to includes to
have single source of truth/text for that.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>