docs: do not jump to the conclusion in fix dry-run
The search gives mostly titles and a small subset of the text.
But our current header does not describe what we achieve, it describes
what we use to do so.

But that information "use dry-run" is what we expect our readers to
learn from reading this page. Therefore change it from "what we use"
to "what we achieve" to improve the TOC overview and search results.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>