docs: the new fix scenario is an explanation
With the new split around the pro fix command, the list of fix scenarios
is actually neither a how-to not a tutorial anymore, it is an
explanation and therefore we will order it that way.

The formerly established cross-links among the related topics help a lot
to still be able to consume this as one large topic despite crossing
three diataxis quadrants now.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>