debian: use --no-block when starting ubuntu-advantage.service
ubuntu-advantage.service is "After: cloud-config.service".
--no-block ensures that it doesn't cause deadlock when cloud-init tries
to install ubuntu-advantage-tools.

This is accomplished by no longer using dh_systemd_start, but taking the
snippets it used to add to maintscripts and hardcoding them. Then I
added --no-block to the "start" command.

I took the latest version of the dh_systemd_start snippets for prerm,
postrm, and postinst. Those snippets are slightly different from those
automatically added on older releases, namely xenial and bionic. The
primary differences are: (1) the addition of checks to only run the snippets
on appropriate dpkg operations by checking $1, and (2) checking that
DPKG_ROOT is empty in addition to the presence of /run/systemd/system
before running systemctl commands.