maintscripts: mv_conffile all conffiles in old package
This commit uses mv_conffile on all of the conffiles in the last version
of ubuntu-advantage-tools. It does not actually move them (the source and
destination are the same), but this removes the "obsolete" conffile
status from the ubuntu-advantage-tools package after upgrade. All of
these conffiles are listed correctly for ubuntu-pro-client after
upgrade.

The exception is if a user had modified one of these conffiles. In that
case, the changes are preserved, but the file is still listed as
"obsolete" for the ubuntu-advantage-tools package after upgrade.