remove machine token from UAConfig
Our goal is to make the UAConfig module only handle user configuration.
Right now it also handles any interaction that we have with the machine
token file. We are removing that responsability from the UAConfig class
by allowing the necessary classes and functions to access the machine
token file directly