docs: release 31.2 (#2988)
* Docs apparmor update (#2906)

* doc: Add security features troubleshooting doc

This adds a document explaining how to troubleshoot the Apparmor and
systemd sandboxing features added to the apt-news service.

* global replace of apparmor with AppArmor

* Review from Sally: wording on the description of how applications can behave under confinement

Co-authored-by: Sally <sally.makin@canonical.com>

* Systemd capitalization

* Turns out it should be systemd instead (lowercase)

---------

Co-authored-by: Sally <sally.makin@canonical.com>

* docs: add guide to setup windows machine for WSL

Add guide to setup windows machine for running our
integration tests on WSL instances

* Update fix docs (#2931)

* docs: update pro fix examples

The pro fix commands were updated recently. Due to that,
we need to update the output examples in the docs.

* docs: better visualize fixing multiple CVEs

Add a how-to-guide to allow users to better visualize the status of each
CVE they have try to fix when running the fix execute API endpoint

* remove unnecessary anchor that linkcheck doesn't like

* Apply suggestions from code review

Co-authored-by: Sally <sally.makin@canonical.com>

* Update wordlist.txt

There's a jq in a link as well, so spellcheck still fails. I've re-added jq to the wordlist.

---------

Co-authored-by: Andreas Hasenack <andreas@canonical.com>
Co-authored-by: Sally <sally.makin@canonical.com>
Co-authored-by: Lucas Moura <lucas.moura@canonical.com>