apt-news.service: ignore apparmor errors when starting
It's possible to have a system where apparmor is enabled, but no profile
is loaded. This happens when the user removed the apparmor debian
package. In that situation, the ubuntu_pro_apt_news profile will not be
loaded into the kernel at startup, and systemd will fail to apply it to
the service when starting.

Systemd already ignores the apparmor setting when apparmor is disabled,
but in this situation it is still enabled. The fix is to configure the
service to also ignore errors when apparmor is enabled.

The drawback is that the service might run unconfined now if there are
problems applying the selected apparmor profile on startup.

Fixes: #3002
LP: #2057937