logs: add redundancy to secret redaction
This introduces an additional method of ensuring secrets don't end up in logs.

This adds known secrets that are in memory anyway to a centralized list and checks every log string if it contains one of those secrets before logging it. If it contains a secret, it gets redacted.

This is less fragile than the regex-based redaction method already in place, but this has the gap of potential logging before a secret gets added to the manager. Taken together, we can be a bit more confident in secret redaction overall.