gpg: create new key for esm-infra-legacy service
Since esm-infra-legacy and esm-infra are independent services, we want
to avoid as much as we can for one service to interfere with the other.
However, since both services share the same gpg key, if the user
disables esm-infra-legacy, we would also affect esm-infra, as the
gpg key would be deleted and apt update would alert the user about
that. To fix that issue, we are now creating a copy for the esm-infra
gpg key for the esm-infra-legacy service. This will guarantee that
each service will only act upon their respective keys during a disable
operation