tests: fix APT news tests for bionic
On Bionic, the distro-info-data package version
is bigger than what we have on ESM. Therefore,
if we upgrade when attached, we will get the ESM version,
due to the pin file. After we detach and run apt upgrade,
distro-info-data will then be upgraded. Since our tests
required that we have no packages to upgrade when checking the
apt news messages, we are now adding an additional apt upgrade
step after we detach just to get rid of this issue.